test: factory ratchetのscope解決を補強
lambda引数と各comprehension targetを独立scopeとして扱い、module model aliasのshadowを誤計上しない。解析責務を専用test helperへ分割する。

検証: offline 2012 passed, 4 skipped; ratchet/ruff/mypy passed
